refactor : add velodyne actor  - Reorganize the 'LidarBaseActor' to include only the common elements of the sensor
